[Core] Fixed header guard for [Serial] mode, removed more Linux compilation warnings

+++ include/occaParserTypes.hpp
@@ -423,7 +423,7 @@ namespace occa {
     void removeArg(const int pos);
 
     inline bool argIsConst(const int pos) const {
-      if(pos < argumentInfos.size())
+      if(((size_t) pos) < argumentInfos.size())
         return argumentInfos[pos].isConst;
 
       return false;
